Fix duplicate report logic and move files into correct project structure
